requirements.txt
PyYAML==6.0
Jinja2==3.1.2
resotolib>=3.4.1
